Prod-0: Move JPA factories to dal package
